requirements.txt
streamlit
cohere
faiss-cpu
python-dotenv
numpy
